Смена дизайна всех диалогов

+++ app/src/main/java/com/skyfolk/quantoflife/ui/create_quant/CreateQuantDialogFragment.kt
@@ -16,6 +16,7 @@ import com.skyfolk.quantoflife.R
 import com.skyfolk.quantoflife.databinding.CreateQuantDialogBinding
 import com.skyfolk.quantoflife.entity.*
 import com.skyfolk.quantoflife.settings.SettingsInteractor
+import com.skyfolk.quantoflife.utils.showConfirmDialog
 import java.lang.reflect.Field
 
 class CreateQuantDialogFragment(
@@ -123,7 +124,18 @@ class CreateQuantDialogFragment(
             }
 
         binding.buttonDelete.setOnClickListener {
-            dialogListener?.onDelete(quant!!)
+            requireContext().showConfirmDialog(
+                title = resources.getString(R.string.delete_quant_title),
+                message = resources.getString(R.string.delete_quant_message),
+                positiveButtonTitle = resources.getString(R.string.delete),
+                negativeButtonTitle = resources.getString(R.string.cancel)
+            ) {
+                dialogListener?.onDelete(quant!!)
+                dismiss()
+            }
+        }
+
+        binding.buttonBack.setOnClickListener {
             dismiss()
         }
